feat: Change reminder_hours_before type from int to float in ReminderConfigUpdate

+++ routes/business_settings_routes.py
@@ -54,7 +54,7 @@ def update_fcm_token(data: FcmTokenUpdate, user_id: str = Depends(obtener_usuari
 
 class ReminderConfigUpdate(BaseModel):
     business_id: str
-    reminder_hours_before: int
+    reminder_hours_before: float
 
 
 @router.get("/business-settings")
